feat(cli): Rename dataset backup subcommand to just backup, handle file names that contain a path segment to prevent archiving failure
