+++ packages/sanity/src/_internal/cli/commands/backup/disableBackupCommand.ts
@@ -1,17 +1,17 @@
 import {CliCommandDefinition} from '@sanity/cli'
-import resolveApiClient from '../../../actions/dataset/backup/resolveApiClient'
-import {defaultApiVersion} from './datasetBackupGroup'
+import resolveApiClient from '../../actions/backup/resolveApiClient'
+import {defaultApiVersion} from './backupGroup'
 
 const helpText = `
 Examples
-  sanity dataset-backup enable <dataset-name>
+  sanity backup disable DATASET_NAME
 `
 
-const enableDatasetBackupCommand: CliCommandDefinition = {
-  name: 'enable',
-  group: 'dataset-backup',
+const disableDatasetBackupCommand: CliCommandDefinition = {
+  name: 'disable',
+  group: 'backup',
   signature: '[DATASET_NAME]',
-  description: 'Enable backup for a dataset.',
+  description: 'Disable backup for a dataset.',
   helpText,
   action: async (args, context) => {
     const {output, chalk} = context
@@ -28,16 +28,17 @@ const enableDatasetBackupCommand: CliCommandDefinition = {
         headers: {Authorization: `Bearer ${token}`},
         uri: `/projects/${projectId}/datasets/${datasetName}/settings/backups`,
         body: {
-          enabled: true,
+          enabled: false,
         },
       })
-      output.print(`${chalk.green(`Dataset backup enabled\n`)}`)
+      output.print(`${chalk.green(`Disabled daily backups for dataset ${datasetName}\n`)}`)
     } catch (error) {
       const msg = error.statusCode
         ? error.response.body.message
         : error.message || error.statusMessage
-      output.print(`${chalk.red(`Enabling dataset backup failed: ${msg}`)}\n`)
+      output.print(`${chalk.red(`Disabling dataset backup failed: ${msg}`)}\n`)
     }
   },
 }
-export default enableDatasetBackupCommand
+
+export default disableDatasetBackupCommand
